Revert "Fix various signed/unsigned warnings"
It appears at least on of those casts changes the semantics

This reverts commit cf254f9a9a04a94ba673b349fa864d6c043366d9.

# Conflicts:
#	src/EHHeaderParser.hpp

+++ libunwind/src/DwarfInstructions.hpp
@@ -71,7 +71,7 @@ private:
       // there's a single register for the stack.
       return registers.getRegister(UNW_MIPS_C11) + 
           registers.getRegister((int)prolog.cfaRegister) +
-             (uint64_t)prolog.cfaRegisterOffset;
+             prolog.cfaRegisterOffset;
 #else
       return (pint_t)((sint_t)registers.getRegister((int)prolog.cfaRegister) +
              prolog.cfaRegisterOffset);
